+++ frontend/src/app/pages/exercises/exercise/shared/simulation/signaller-modal/details-modal/eoc-information/signaller-modal-eoc-information-alarm-groups-sent/signaller-modal-eoc-information-alarm-groups-sent.component.ts
@@ -22,7 +22,7 @@ export class SignallerModalEocInformationAlarmGroupsSentComponent {
         this.alarmGroupsSent = Object.values(
             selectStateSnapshot(selectAlarmGroups, store)
         )
-            .filter((alarmGroup) => alarmGroup.sent)
+            .filter((alarmGroup) => alarmGroup.triggerCount > 0)
             .map((alarmGroup) => alarmGroup.name)
             .sort((a, b) => a.localeCompare(b));
     }